+++ yash-env/src/system/signal.rs
@@ -392,11 +392,17 @@ pub trait Sigmask: Signals {
     /// applying the given `SigmaskOp` and signal set to the current mask. If
     /// `op` is `None`, this function does not change the mask.
     /// If `old_mask` is `Some`, this function sets the previous mask to it.
+    ///
+    /// The return type is a `Future` to support simulating blocking I/O in
+    /// [virtual systems](super::virtual). In the [real system](super::real),
+    /// this function does not work asynchronously and returns a ready `Future`
+    /// with the result of the underlying system call. See the
+    /// [module-level documentation](super) for details.
     fn sigmask(
         &self,
         op: Option<(SigmaskOp, &[Number])>,
         old_mask: Option<&mut Vec<Number>>,
-    ) -> Result<()>;
+    ) -> impl Future<Output = Result<()>> + use<Self>;
 }
 
 /// How the shell process responds to a signal